fix(work): move telemetry to work page
