+++ src/app/page.tsx
@@ -1,4 +1,3 @@
-import { Suspense } from "react";
 import { headers } from "next/headers";
 import Image from "next/image";
 import Link from "next/link";
@@ -11,39 +10,10 @@ import {
   experienceEntries,
 } from "@/content/site";
 import { InteractiveTerminal } from "@/components/home/InteractiveTerminal";
-import { GitHubCommitPulse } from "@/components/home/GitHubCommitPulse";
 import { TypingTestGate } from "@/components/home/TypingTestGate";
 import { Reveal } from "@/components/motion/Reveal";
 import { CountUpMetric } from "@/components/motion/CountUpMetric";
 import { TypeOnReveal } from "@/components/motion/TypeOnReveal";
-
-function GitHubCommitPulseSkeleton() {
-  return (
-    <>
-      <style>{`
-        .pulse-skeleton {
-          background: var(--color-surface);
-          animation: pulse-skeleton-fade 1.8s ease-in-out infinite;
-        }
-
-        @keyframes pulse-skeleton-fade {
-          0%, 100% { opacity: 0.55; }
-          50% { opacity: 0.9; }
-        }
-      `}</style>
-      <div
-        className="pulse-skeleton"
-        aria-hidden="true"
-        style={{
-          padding: "var(--space-8)",
-          minHeight: "20rem",
-          border: "1px solid var(--color-border)",
-          borderRadius: "var(--radius-lg)",
-        }}
-      />
-    </>
-  );
-}
 
 export default async function Home() {
   const nonce = (await headers()).get("x-nonce") ?? undefined;
@@ -360,16 +330,6 @@ export default async function Home() {
           ~/work &rarr; See the case studies
         </Link>
       </div>
-
-      <Reveal>
-        <section style={{ marginBottom: "var(--space-12)" }}>
-          <div className="container">
-            <Suspense fallback={<GitHubCommitPulseSkeleton />}>
-              <GitHubCommitPulse />
-            </Suspense>
-          </div>
-        </section>
-      </Reveal>
 
       <TypingTestGate />
