+++ src/app/work/work.module.css
@@ -375,3 +375,7 @@
   color: var(--color-accent);
   transform: translateX(4px);
 }
+
+.telemetrySection {
+  margin-top: var(--space-12);
+}
